passenger-bot/requirements.txt
aiogram==3.4
python-dotenv
googlemaps==4.10.0
pydantic<2.0
folium
geopy
tzdata